+++ src/V1/Checks/IsShortColorValue.php
@@ -1,0 +1,79 @@
+<?php
+
+/**
+ * Copyright (c) 2017-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   CssParser\Checks
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://code.ganbarodigital.com/php-mv-css-parser
+ */
+
+namespace GanbaroDigital\CssParser\V1\Checks;
+
+use GanbaroDigital\CssParser\V1\Values\CssColorValue;
+
+/**
+ * can the color be represented using #RGB shorthand notation?
+ */
+class IsShortColorValue
+{
+    /**
+     * can the color be represented using #RGB shorthand notation?
+     *
+     * @param  CssColorValue $color
+     *         the CssColor to examine
+     * @return bool
+     *         TRUE if we can render the color using #RGB
+     *         FALSE otherwise
+     */
+    public static function check(CssColorValue $color)
+    {
+        list($red, $green, $blue, $alpha) = $color->getSRGB();
+
+        // #RGB notation does not support alpha channels
+        if ($alpha !== null) {
+            return false;
+        }
+
+        // each constituent must be divisible by 17 for this to work
+        if ($red % 17 === 0 && $green % 17 === 0 && $blue % 17 === 0) {
+            return true;
+        }
+
+        // if we get here, we've run out of tests to attempt
+        return false;
+    }
+}